# docs/api.yaml
openapi: 3.0.0
info:
  title: WISEducation API
  description: API contract for WISEducation platform covering Auth, Market, Education, Community, and Simulation.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: string
        reputationPoints:
          type: integer
...
    Post:
      type: object
      properties:
        id:
          type: string
        author:
          $ref: '#/components/schemas/User'
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        type:
          type: string
          enum: [text, image, poll]
        upvotes:
          type: integer
        commentCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        pollOptions:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              votes:
                type: integer
        imageUrl:
          type: string
...
paths:
...
  /community/posts:
    get:
      summary: Get community feed
      parameters:
        - in: query
          name: sort
          schema:
            type: string
            enum: [trending, new, top]
        - in: query
          name: filter
          schema:
            type: string
      responses:
        '200':
          description: Feed posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: Create a new post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                type:
                  type: string
                  enum: [text, image, poll]
                imageUrl:
                  type: string
      responses:
        '201':
          description: Post created
...